SQL - add --postgres-v2 option to generate postgres columns with jsonb
Duplicate default postgresDbProfile to newer postgresDbProfileV2

postgresDbProfileV2 to store unmapped ADL types in postgres as jsonb instead of json

Since postgres 9.4 but especially since 9.6 it has had good support for jsonb
Plain json column stores in the DB as a string and only basically parses out the json when needed.
Jsonb column stores it in the DB as a parsed and binary - postgres has much better capabilities on jsonb than json columns.

https://www.compose.com/articles/faster-operations-with-the-jsonb-data-type-in-postgresql/
https://www.postgresql.org/docs/9.6/functions-json.html